0.1.0 version of Kindle notes parser
`README.md` file explaining purpose of this script and its usage is added.
Some ideas about what to do next and what had to be done first, too.

Base structure of the repository was established. It's sub-optimal, but
for now, the script was meant to work at least on provided example of
`My Clippings.txt` file as downloaded from `Kindle Paperwhite 4`.

Base logic was implemented to read input file and create markdown files
in target directory. Its performance is not optimal but will be performed
in future releases.

At this point, even the name of the script is not really established.
For working purposes, `kindleparse` is used, but it may change in the
nearest future.

Some unit tests covering critical functionality were implemented.

Ruff was used for linting and formatting, though at this point it is not
used as part of CI.

A lot of mockery was introduced to parse text from Kindle. Now it works,
but is not "clean".

Minimal CLI is provided using Python's `argv`. Some fancier solution
may be implemented later.

The commit history was squashed into this one as it was ugly and made
no sense at this point.